Reduce abstraction overhead in color conversion
Previously, ImageDecoderPdn had been:
  1. Converting the PSD channel-separated format to System.Drawing.Color
  2. Converting the Color to a PaintDotNet.ColorBgra
  3. Copying the ColorBgra into the PaintDotNet.MemoryBlock, one byte
     at a time.

This change collapses the three copies into one, by changing the
GetColor functions into SetColor functions that operate directly on the
MemoryBlock.

This results in a 2x performance improvement for PSD loading.  Most of
the time is now spent in MemoryBlock, which is Paint.NET code.